requirements.txt
flask==2.0.1
pymongo==4.13.0
python-dotenv==1.1.0
